Refactored CompiledSDFG and invocation to its own file
